migrate TrsTool to vue composition api
